Moved checkCapsLock function to utils, removed validatInput
