+++ src/components/SignUp.jsx
@@ -2,6 +2,7 @@ import { useState, useRef } from "react";
 import { registerUser } from "../utils/auth";
 import { Navigate } from "react-router-dom";
 import Loading from "./Loading";
+import useCapsLockCheck from "../utils/checkCapsLock";
 import "../styles/signUp.css";
 
 // ToDo: Add Input Fields for business customer
@@ -65,11 +66,10 @@ const SignUp = ({
     tel: "",
   });
   const errorRef = useRef();
-  const [isCapsLockOn, setIsCapsLockOn] = useState(false);
+  const [isCapsLockOn, checkCapsLock] = useCapsLockCheck();
 
   const handleChange = (e) => {
     setFormState((prev) => ({ ...prev, [e.target.id]: e.target.value }));
-    validateInput(e);
   };
 
   // Submit the form
@@ -193,110 +193,6 @@ const SignUp = ({
     }
   };
 
-  // Validate the input
-  const validateInput = (e) => {
-    let { id, value } = e.target;
-    setError((prev) => {
-      const stateObj = { ...prev, [id]: "" };
-
-      switch (id) {
-        case "email":
-          if (!value) {
-            stateObj[id] = "Bitte Email angeben.";
-          }
-          break;
-
-        case "password":
-          if (!value) {
-            stateObj[id] = "Bitte Passwort angeben.";
-          } else if (confirm_password && value !== confirm_password) {
-            stateObj["confirm_password"] = "Passwort stimmt nicht überein.";
-          } else {
-            stateObj["confirm_password"] = confirm_password
-              ? ""
-              : error.confirm_password;
-          }
-          break;
-
-        case "confirm_password":
-          if (!value) {
-            stateObj[id] = "Bitte Passwort wiederholen.";
-          } else if (password && value !== password) {
-            stateObj[id] = "Passwort stimmt nicht überein.";
-          }
-          break;
-
-        case "salutation":
-          if (!value) {
-            stateObj[id] = "Bitte Anrede angeben.";
-          }
-          break;
-
-        case "firstname":
-          if (!value) {
-            stateObj[id] = "Bitte Vornamen angeben.";
-          }
-          break;
-
-        case "lastname":
-          if (!value) {
-            stateObj[id] = "Bitte Nachnamen angeben.";
-          }
-          break;
-
-        case "birth_date":
-          if (!value) {
-            stateObj[id] = "Bitte Geburtsdatum angeben.";
-          }
-          break;
-
-        case "zip_code":
-          if (!value) {
-            stateObj[id] = "Bitte Postleitzahl angeben.";
-          }
-          break;
-
-        case "city":
-          if (!value) {
-            stateObj[id] = "Bitte Stadt angeben.";
-          }
-          break;
-
-        case "street":
-          if (!value) {
-            stateObj[id] = "Bitte Straße angeben.";
-          }
-          break;
-
-        case "street_number":
-          if (!value) {
-            stateObj[id] = "Bitte Hausnummer angeben.";
-          }
-          break;
-
-        case "country":
-          if (!value) {
-            stateObj[id] = "Bitte Land angeben.";
-          }
-          break;
-
-        default:
-          break;
-      }
-
-      return stateObj;
-    });
-  };
-
-  // Check if the caps lock is on
-  const checkCapsLock = (event) => {
-    if (event.getModifierState("CapsLock")) {
-      setIsCapsLockOn(true);
-    } else {
-      setIsCapsLockOn(false);
-    }
-  };
-
   if (loadingAuthRequest) return <Loading />;
   if (isAuthenticated) return <Navigate to="/auth" />;
 
@@ -322,7 +218,6 @@ const SignUp = ({
               type="email"
               placeholder="E-Mail Adresse"
               value={email}
-              onBlur={validateInput}
               onChange={handleChange}
             />
 
@@ -341,7 +236,6 @@ const SignUp = ({
               placeholder="Passwort"
               value={password}
               onChange={handleChange}
-              onBlur={validateInput}
               onKeyUp={checkCapsLock}
             />
             {error.password && (
@@ -360,7 +254,6 @@ const SignUp = ({
               placeholder="Passwort wiederholen"
               value={confirm_password}
               onChange={handleChange}
-              onBlur={validateInput}
               onKeyUp={checkCapsLock}
             />
             {error.confirm_password && (
@@ -384,7 +277,6 @@ const SignUp = ({
               id="salutation"
               value={salutation}
               onChange={handleChange}
-              onBlur={validateInput}
             >
               <option value="">- Bitte Anrede wählen -</option>
               <option value="Frau">Frau</option>
@@ -406,7 +298,6 @@ const SignUp = ({
               placeholder="Vorname"
               value={firstname.charAt(0).toUpperCase() + firstname.slice(1)}
               onChange={handleChange}
-              onBlur={validateInput}
             />
             {error.firstname && (
               <small className="err" ref={errorRef}>
@@ -423,7 +314,6 @@ const SignUp = ({
               placeholder="Nachname"
               value={lastname.charAt(0).toUpperCase() + lastname.slice(1)}
               onChange={handleChange}
-              onBlur={validateInput}
             />
             {error.lastname && (
               <small className="err" ref={errorRef}>
@@ -441,7 +331,6 @@ const SignUp = ({
                 placeholder="Straße"
                 value={street.charAt(0).toUpperCase() + street.slice(1)}
                 onChange={handleChange}
-                onBlur={validateInput}
               />
 
               <input
@@ -451,7 +340,6 @@ const SignUp = ({
                 placeholder="Nr./ID"
                 value={street_number}
                 onChange={handleChange}
-                onBlur={validateInput}
               />
             </div>
             {error.street && (
@@ -477,7 +365,6 @@ const SignUp = ({
                 placeholder="PLZ"
                 value={zip_code}
                 onChange={handleChange}
-                onBlur={validateInput}
               />
 
               <input
@@ -487,7 +374,6 @@ const SignUp = ({
                 placeholder="Stadt"
                 value={city.charAt(0).toUpperCase() + city.slice(1)}
                 onChange={handleChange}
-                onBlur={validateInput}
               />
             </div>
             {error.zip_code && (
@@ -508,7 +394,6 @@ const SignUp = ({
               id="country"
               value={country}
               onChange={handleChange}
-              onBlur={validateInput}
             >
               <option value="">- Bitte Land wählen -</option>
               <option value="Deutschland">Deutschland</option>
@@ -538,7 +423,6 @@ const SignUp = ({
               placeholder="DD.MM.YYYY"
               value={birth_date}
               onChange={handleChange}
-              onBlur={validateInput}
             />
             {error.birth_date && (
               <small className="err" ref={errorRef}>
